Renaming in vocab resolution; finding bugs
(1) It was concerning that something as important as renaming wasn't implemented in vocab resolution but so many tests were still succeeding.  It took me quite a lot of mental effort to just find a test that failed, and then looking at the failure revealed our push algorithm was doing something really stupid, it was replacing *every* ref rather than replacing the correct ref!

How did so many tests still pass?  Well, for a bunch of reasons, but either (a) they were match tests that had the same vocab anyway, and (b) MatchExpr no longer checks whether the opposite-tape vocab contains the buffer literal anyway, so it can work even when the vocab is wrong, and (c) for tapes bound to explicit cursors we can no longer check the contents of vocab.  What I needed to do to make a failure: make a match where the vocabs HAD to differ and then put a dot that reveals in the outputs what the vocab was.

(2) It's clearer now why a variety of bugs all make the same match/replace tests "fail" in the same way, and in the direction of "outputs that WILL eventually be correct".  Auto-priority puts Cursors on grammars in a predictable way, and so bugs that mistake one tape for another will tend to get the same effect on a ton of examples -- usually using the output vocab for the input.  This result happens to be what we want in the future, but it's still evidence of a bug in the present.

(3) I still haven't quite figured out the right vocab resolution algorithm, fixing our "push" algorithm to avoid the bug in (1) reveals it's still not managing to "scrub" all the refs out.  Cursor t1 doesn't just need to scrub t1 from its *own* tapes, but also from the tapes of its children, or else a child cursor t2 could still have a t1 wildcard on one of its tapes.  So we really do need something like the env.vocabMap of the "pull" algorithm to handle this.