Merge keys operation
It became clear that the *old* new vocab system (call it the "Wildcard System") had issues in practice.  We invented that system in order to enable the change in Cursor/tape semantics, but in the end they made some incompatible assumptions.

We were handling inter-tape relationships (e.g. t1 being a superset of t2) with wildcard referencees (e.g. t1 containing as part of its vocab a `Ref(t2)` referring to whatever is in t2's vocab).  However, once the drastic change in Cursor semantics actually went down, some of these wildcards ended up referring to tapes that don't, in that scope, exist.  It's possible that there was some clever solution where we push some info into the env, get other information from returns, etc., but actually ordering that solution was dicey and in practice would require more recalculation passes than I was happy with.

So we're scrapping that in favor of something that (in some ways) is simpler and more restrictive, but other ways a bit more complex.  Instead of handling the equivalence of t1 and t2 with wildcard refs in each other's vocabs, we merge the vocabs under a new name chosen at random, and make both t1 and t2 point to that.

(Eventually we want to make it so this name isn't random, but having stable global names of things has been buggy in the past.  So for now it's random and I'll deal with it later.)

This is complicated by the fact that the input vocab dicts might themselves contain references from previous merge operations, and any operations on those dicts have to handle that appropriately.

This commit just concentrates on the merge operation itself.  Next we need a sum/product/intersection operation that also follows references correctly, and finally we'll need a "scrub" operation that eliminates all references to a given tape (without this, names could escape their Cursors and lead to encapsulation failures).