Preparing for npm publishing the cli and interpreter packages.
Added LICENSE and PRIVATE.md files to the cli, interpreter, sheets_addon packages.
Added references to these files to the README files for these packages. Also added
copyright notices and reference to the Gramble language documentation to these
README files.

Added a brief package description to the README file for the interpreter, but it
still needed usages documentation.

Set the version number as 0.9.0, so we can do a trial npm publish to iron out any
difficulties before we try to publish version 1.0.0.