Getting VocabDict sums working
This is a sum operation on new-style VocabDicts where keys can potentially be Refs to other keys, e.g.

   {t1: X, X: {a,b}} + {t1: Y, Y: {a,c}}

The result needs to have three properties:

   (1) all of the keys of the originals still exist, even intermediate keys like "X" and "Y"
   (2) all keys form chains leading eventually to literal vocabs
   (3) if at any point a key is shared between both input sets, the key in the output must refer to the sum of its eventual referent vocabs.

So for the above, the following answer would be valid:

   {t1: X, X: Y, Y: {a,b,c}}

We could also introduce new keys, like we did in mergeKeys, we just can't remove any keys.  And we do that, actually, because we call mergeKeys here as a convenient way to fix up certain kinds of chains.  So really the result looks like:

   {t1: X, X: Y, Y: $9ym09eru8l, $9ym09eru8l: {a,b,c}}

In the last commit, I was slightly on the wrong track.  I was taking the general idea from mergeKeys() that, when we hit a reference, we had to follow that reference (that is, recurse using the reference's key in place of the original key).  But in doing this in sumKey(), I was getting into various kinds of infinite recursions, all for somewhat silly reasons that I won't go into because they're all moot now.

Ultimately, my conceptual mistake was that I was doing ref following/repair on BOTH sides of sumKey() -- that is, both references that occur in dict1, which we use as a sort of accumulator, and dict2, which is the dict we're looping through and summing key-wise with dict1.  (mergeKeys(), in contrast, only has one dict, the accumulator.)  Following/fixing refs in dict2 is both dangerous (it sets up some possible infinite recursions) and also unnecessary (any keys in dict2 will eventually be copied over to dict1 because we're looping through them in sumKeys()).  Really the only thing we need to do is a little ref-chain fixup in dict1, anticipating that the ref in question is eventually going to be copied over and needs to be part of a valid chain, and not obliterate any literal content in the process.  (Or really, we don't know what order these operations will occur, the ref may ALREADY have been copied over and here we're fixing things up to account for that.  It works either way.)

We could implement this fixup mechanism in various ways, but luckily one such fixup is identical to mergeKeys.  (It's not the absolute most minimal fixup, because mergeKeys adds an additional key that isn't absolutely necessary for this fixup, but I'd rather reuse that code than have two similar algorithms doing very similar things.)