Making new vocabs work with atomicity
Although the new Lits had atomicity (because they were just copy-pasted over from the old Lits), we weren't actually using/testing it; everything up to this point had been sums of Tokenizeds.

This commit generalizes sumKeys as a sort of map function, which can take any of {sum, product, intersect} and apply it correctly to new-style VocabDicts.

I added a lot of tests of this, and redid the testing API somewhat.  Previously I allowed vocabs like `{t1: ["a","b"]}`, and defaulted to them being interpreted as Tokenized, but now we had need of every kind of Lit, and it was important to specify which kind.  I found myself getting confused at different points of the code's evolution about what these represented -- atomics, tokenized, ...?  If I say ["ab"], is that treated like a LiteralGrammar and tokenized, or is that one "ab" token?

So I nixed that, and the whole vocToVocabDict convenience function, and instead made `Atom(...ss)`, `Conc(...ss)`, and `Tok(...s)` convenience functions instead.  You have to be explicit about which of these you intend, but because they use `...` params it's not really that much more typing than `["a","b"]`.  `Tok` *does* tokenize like we would a LiteralGrammar, `Tok("ab")` ends up with the vocab `{"a","b"}`.