Reattaching vocab system to tape system
Replacing the old vocabs with the Tape/TapeCalc system to be the new vocabs.

(1) Alphabet.ts was just a placeholder file, so I moved all the code from alphabet.ts into vocab.ts and deleted the former.

(2) Unlike before, the set of keys in the vocabMap is not necessarily the set of tapes!  There's no tape "$ai49xv8sq", that's just a sort of breadcrumb we've left in the vocab map for later.  So we have to keep the set of tape names separately.

(3) Most of the tape operators changed, largely for the simpler.  Most of the tapeCalc didn't because most of that didn't touch vocabs directly, the only exception is when we had to place wildcards for particular kinds of grammars.  Now we don't have wildcards but still have to make sure appropriate tapes are Tokenized for those.

(4) There are still some failing testPassTapes tests, mostly because vocab results have actually changed.  (E.g., replace cascades are no longer supersets, they're equivalent.)  But I'm not sure that's the only problem; I just wanted to commit this before I changed anything further.