Fixing up failed tape tests
(1) The vocabs of replace inputs/outputs has changed, so most of the replace tests failed.  Something like replacing `e -> a` in "hello" used to have an input vocab of [h,e,l,o] and an output vocab of [h,e,l,o,a].  But talking with the programmers, they found that unintuitive that a "tape" (from their point of view) would have different vocabularies like that.  So we decided they would all be the same -- the vocab would be [h,e,l,o,a] both before AND after the replace operation -- and the new vocab system does that.

So all but one of these fixes is just changing the test to the new correct answer.

(2) The other failed test was actually a bug.

When `mapKeys(dict, key, newValue, op)` didn't find that key in the dict, it defaults to an empty Atomic.  (This is just for convenience, doing so is easy to express in JS with || and saved us an `if` block.)  This was fine when we were only doing sums.  But doing so actually gets the wrong answer for products: Atomic & Atomic == Concatenated!  So by putting in an atomic default, we were actually promoting some tapes that should have been Atomic to Concatenated, and then to Tokenized.