Hooking up new vocabs to whole system
(1) The new vocab system works, and we're green on everything that was giving us trouble (e.g. Cursors, Matches, etc.).

(2) There's a new pass, ResolveVocab, that comes right before constructExpr.  It takes the vocabMap at the root (which at this point contains NO tape names, but contains all the intermediate references) and passes it through the tree, allowing Cursors that don't already have literal vocabs to follow the final chain of references to their vocabs.

(We don't want to follow these chains early, like during tapeCalcing CursorGrammars, even though it seems like the vocab is available then.  It's possible that merges outside of the scope of the CursorGrammar change the vocab after the Cursor has been tapeCalced.  It's only at the very end that we can nail down all vocabs for certain.)

(3) Running unit tests was once again a sticking point for this.  We have to ResolveVocab right before the initial constructExpr call in runTests() -- the one that only exists to populate the symbol table -- because otherwise any explicit Cursors who are involved in Matches won't receive their vocabs.

Unit test running remains a common source of bugs for anything late in the compilation process, simply because it runs so strangely -- it's doing executions way up the tree.  We should think of a better way to do this, like having the tests create hidden identifiers and then calling an Embed of those near the top of the tree.

(4) Last time, we split Tapes.Lit into having both a tapeNames member for the actual tapes and vocabMap with a map to their vocabs (and potentially having keys that aren't tapes).  There's one grammar where we didn't quite get this right during TapeCalc, CorrespondGrammar.  (CorrespondGrammar can add a tape name that doesn't exist in its child -- like even if you have an epsilon input, which has no tape, you still need the CorrespondGrammar to report that it has a tape "$i".)  We were adding "$i" to the vocabMap but forgot to add it to the tapeNames.

(5) The facts of vocab have changed: matched tapes, including the input and output tapes of replaces, now have identical vocabs.

This means that any test that gave them DIFFERENT vocabs is now broken.  We'll have to go through and decide whether to update them to the correct answers or remove them.