Small safety check in Cursor tapecalc
Cursor.vocab gets its value from the cursor's child's vocabMap.  Then, if that value is a Ref, it's later resolved to a Lit during the ResolveVocab pass.

However, if we were to call tapecalc again (which is very easy to do via tapify), that resolved Lit would be replaced by the old Ref!

We don't do that, but it'd be very easy to forget and inadvertantly do this (like if in the future another pass intervenes in between ResolveVocab and constructExpr, we'd probably call tapify on that and it'd overwrite the Lit).

So now we check to make sure Cursor.vocab isn't already a Lit before overwriting it in tapecalc.