Fixing & refactoring Messages
I had accidentally turned off the ability of the GSuite add-on to respond properly to Messages (i.e., syntax highlighting, syntax errors/warnings, unit test results).  I had renamed the "type" member of Message to "tag" so that it was like other kinds of components in our system, but the add-on was still expecting "type".  (The add-on files are vanilla JavaScript so we didn't get any type errors.)

While I was at it, I refactored Messages and some of the code that handles them, so that they're more like other components.  (E.g., instead of passing in a "type" string to the constructor of each, the subclasses define a readonly tag, and the strings are defined in a const enum.)

The reason this was such a heterogeneous mess before was that I didn't understand exactly about TypeScript objects would survive transpilation to JavaScript/GAS.  (There were always some surprises, like GAS not allowing inheritance between objects defined in different files.) So for the APIs that would be called from JS, I was being maximally safe and basically just treating these as a sort of array.  Now I understand better what the relationship is between TS/JS/GAS, and also we use less inheritance to begin with now, so I feel more confident just accepting messages as Message objects.