More Messages fixes/refactoring
(A) Overall, we had about 4 different ways Messages were expressed in the code:

    (1) As Message objects, which got refactored last time.

    (2) As destructured params passed around inside DevEnvironments, which we eliminated last time in favor of just passing around (1)

    (3a) As plain arrays (i.e. with number keys) like ["filename", 4, 3, shortMsg, longMsg, "error"] stored inside SimpleDevEnvironment and SidebarDevEnvironment.

    (3b) As a slightly simpler plain array, returned from getError-type methods in DevEnvironment, only used during testing.

    (3c) As an even simpler plain array, used when creating source tests to say "There should be an error/warning at this position".

This commit makes (3a) and (3b) just use Message objects, and (3c) use Partial<Message>.

(B) The DevEnvironment interface had a lot of different ways to ask "What errors are there?", with different ways of specifying the kind of errors you want and what you wanted to know about them (e.g., an array of fields, a string, or whether you just want how many there are).  These were really only used in testing; different tests called different methods.

Now there's only one, getErrors; it takes a Partial<Message> as a query and returns a list of matching Messages.

SimpleDevEnvironment and SidebarDevEnvironment are also now much simpler inside, because they're not supporting as many ways to slice and dice errors internally.

(C) In source tests, one of the main questions we asked was "Is there a message of a particular type (i.e. error or warning) at this location?"  But we never actually checked whether the error at that location was the right type, because this was inconvenient to do.  Now with the new DevEnv/Message API it's easy to do.

A handful of syntax error checks were of the wrong type (e.g. expecting an error when it should have been a warning), so I went into the tests and fixed those.