Getting outputs from joins and pretapes
We weren't previously able to grab the contents of intermediate tapes before reaching the end of the generation loop (i.e. to print to debug logs), because they were "stuck" inside JoinExpr and PreTapeExpr.  Neither of these had a getOutputs() implementation that allowed us to reach into the parts of the tree where these outputs live.

Now they do, so logDebugOutputs now reports intermediate tape outputs as well.

It's worth noting that the JoinExpr implementation shouldn't be considered "correct" for an arbitrary Expr.  There fundamentally isn't a correct getOutputs result for joins, that's why we're running the generator loop, to simply complex Exprs where we can't straightforwardly know the Output to simple ones where we can.

But for our limited purposes here, we're not trying to get a correct answer, we're simply trying to grab some intermediate Outputs from the tree that are of interest while debugging.  We know where those are -- we built that structure programmatically -- and they're all on the left side of joins, so we can just say "Grab whatever's on the left side of the join and ignore the right".