Preliminary fix for replace-sample garden paths
THE SYMPTOM

I noticed a bizarre slowdown when sampling from grammars with long words and nested rules -- sometimes sampling a single form from a grammar took much more time than generating the ENTIRE grammar.

Digging into the logs, I noticed that the generator loop was managing to create complete entries, and then failing to actually delta out and yield the finished entry.  Instead, it was backtracking -- sometimes backtracking to the very beginning! -- and trying a different route.  Why, and why WASN'T this happening when just generating without randomness?

I speculated in our meeting that we were getting the delta incorrect for some of the weird special expr types in rules, like Correspond or Short, when env.random is true.  But this wasn't the case (we don't actually do anything special with these when they're random).

THE ACTUAL PROBLEM

The issue was a bit deeper and trickier, although obvious in retrospect.  It's that we're getting anything wrong at the level of the derivative, it's the interaction of two things:

(a) The Expr that we construct for rules involves an early split.  You can either search down the path of matching 1+ instances OR search down the path of matching none.  If you're not going to match (and you often won't), going down the path of finding 1+ matches will never succeed, and you won't know it doesn't succeed until the very end.  And this choice of the wrong path can happen at every step on the text field; in order to go down the right path and avoid backtracking you have to flip a coin and get heads every single time.  That's why debug logs show the generator managing to find what looks like a correct complete output and then failing when it tries to delta it, it's the unsimplified material corresponding to the match-pattern that we can't delta out of.

(b) The way we handle alternation during sampling is different than during generation.  In sampling, we're grabbing a single possible next state, and putting that state onto the search stack.  If it turns out we went down a path that will never finish, we'll never go down the correct path until we've thoroughly explored that state and all of its descendants, we have to have popped all of those off the stack until we get to the state right before we made the wrong decision.  That's the backtracking.

This doesn't happen during generation.  Because we explore every next path and combine paths that transitioned along the same character, the correct and incorrect paths don't diverge (because both transitioned along the same character), we're treating them both as one state that retains both possibilities each time.

SOLUTION(S)

Addressing (b) would be quite difficult.  I don't think it's impossible, and maybe it would be useful in the future, but it involves touching the fundamentals of the derivative engine and I don't want to do that so close to the 1.0 release.

So I think we have to address (a).  Now that we're noticing we're doing this, it's no surprise that it's causing efficiency problems, because we've commited the parsing sin of a grammar A|B where the "FIRST()" sets overlap.  They overlap a LOT, both sides start with matchAnythingElse, so basically anything but the target character is going to be an overlap and thus a possibility of going down the wrong path.  (This is no issue in the generation case because disjoin is basically the FIRST() operation.)

So we can address this by rewriting the grammar of rules so that they're LL(1) (or RR(1) when we're going RTL).  That's tricky but in a different way than addressing (b); refactoring what grammar rules produce isn't going to have unexpected consequences elsewhere, whereas touching the generation/deriv fundamentals absolutely will.  And it's something we should do anyway even if someday we do (b).

REFACTORING RULE GRAMMARS

Rewriting the rule grammars in their current form is effectively impossible, though, because we build them in a complex procedure, rather than just say "This kind of rule becomes this formula".

(Why didn't we just specify it like that?  Well, there are a LOT of conditions that affect the grammar of rules.  BeginsWith/EndsWith/Both/Neither, optional/obligatory, minReps at zero or >1, otherContext or none, and whether we're parsing RTL or LTR.  That's 64 possibilities!)

But now we have to bite the bullet; we have to have a grammar be explicit before we can refactor it to be LL(1).

I've made an initial stab at refactoring the grammar-building code so that things are expressed more formulaically, at least enough that I was able to address the early branching problem in normal (non-beginsWith, non-endsWith) rules.  Doing so confirms that yes, this was the problem that was causing slow sampling.

I haven't done the rewriting in general, though; there are plenty of other branching conflicts I haven't addressed yet.  This is just the biggest one and the one programmers are most likely to encounter.