Fix to ConcatExpr laziness
I'm only partway done with the replace rule fixes, but caught a little snag in concatenation that I wanted to address.

Backing up a bit, it seems like my new way of doing replace rules will require a new Expr, because its derivative references itself.  So this requires writing a new Expr class... and that's actually become rather more difficult since the Great Sampling Overhaul.  It's no longer a matter of just yielding everything eventually in whatever order is most convenient, and hope that something higher up makes it random, that's what made sampling slow (because we were generating more than we needed and sampling from that).  Now, instead, we're passing around the derivatives as not-yet-started generators, wrapping them, choosing from lists of generators, etc.  This makes it so we can get reliable and fast sampling, but it's awkward to code, it's sort of halfway between a procedural and analytic way of expressing derivatives.

The new "RewriteExpr" is tricky to express in this sort of half-procedural way.  It'd be easier to just say "It's the alternation of the derivatives of these two expressions".  But both of these contain concatenations, and if you try to construct that concatenation and ask for its deriv, you get into an infinite loop because one child of the concatenation is the RewriteExpr itself, and calling deriv on the concat calls it on that child.  (Rewrites are like a special kind of repeat, both have the same basic "Concat(child, this)" structure.)  But we shouldn't be getting into a loop here, because the child isn't nullable and therefore the nullable-child1 case of Concat shouldn't be firing and calling the deriv of child2.

Turns out that the way we're expressing ConcatExpr.deriv now, deriv gets called on child2 regardless of whether child1 is nullable.  It's not that we don't construct that derivative, it's just that it gets weeded out during simplification.  That's unnecessary in the first place and we're probably wasting some cycles evaluating derivatives that will eventually get simplified to null, so I'm now checking if child1.delta is null before constructing that deriv.

In the big picture, this is just a stop-gap solution, it's clear that writing derivatives in our current way is rather brittle.  I can see a better way to write these now, but it's a potentially big refactoring and getting this rewrite-rule problem sorted out is a higher priority.