Prototype of RewriteExpr
This is a first stab at a more efficient implementation of rewrite rules.

Rewrite rules, both in the old way and in the new way, involve a frequently-repeated decision: do I go down the branch where I'm matching the pattern, or the branch where I'm matching anything BUT the pattern?  This would be easy if patterns were only one character, like "a -> x", because that's unambiguous.  If the character in question is an "a", we go down the match path, and otherwise we go down the anythingElse path.  It's only complicated when the pattern is more complex, like "ab -> x".  If we get an "a", both paths are possible until we get more information later about whether a "b" follows.

The solution here is ultimately rather similar to our old formula, but with a difference in scope.

    * The old solution -- well, the refactoring of the old solution that I did last week -- assembles the formula as "matchAnythingElse | (match + matchAnythingElse)", where matchAnythingElse was a dot-star with a constraint that it not contain instances of the match pattern.  The refactoring eliminates a serious garden path in the older version, but I think there are still subtle garden paths possible here.  (I'm not certain, because it's difficult to think about the character-level derivative of this for anything larger than one character, and looking at the debug traces are difficult because the expressions get so large.  But the fact that those expressions are so complex is concerning in itself.)

    * In the new solution, our matchAnythingElse branch is only ever one character long.  We can go down it multiple times in a row, but that's our only choice at each step: [begin to] match the pattern, or match a dot, and then restart as if we're evaluating a repetition.  (It's not a true RepeatExpr, it's its own expr, but they're similar in this regard.)  Since the pattern can be longer than a single character, however, this wouldn't be compatible with the scoping above (we can't just have a single dot with a constraint that it not match a 2+ character sequence, that would always succeed even in cases where it shouldn't).  Instead, we put a not-begins constraint over that entire branch.  We're saying "Go ahead and go another step down the anythingElse path, but if the results ever end up beginning with <pattern> then null out immediately."