Simplifying joins of dot-stars
The new rewrite rules create some vestigial structure -- for each anythingElse match that succeeds, you end up with a join to a dot-star on the input tape.  (This comes from the not-begins-with constraint, it's the "universe" result of a successful negation.  Once it turns out the anythingElse branch *doesn't* start with the pattern, we're left with $i:.*.)

When a tape T is defined in X, `T:.* ⋈ X` equals X -- that dot-star clause doesn't give any new information.  In general we don't want to create and keep around non-informative structure, because this can result in a growing grammar that kills performance (cf. Might 2011).  It also makes debug logs hard to read.

We hadn't been simplifying this, though, because *in general* this situation can be difficult to detect.  E.g., that T:.* might be nested in something else, like it might be in a sequence `(T1:.* + T2:.*) ⋈ X`.  That *should* simplify to X, but there's no information local to that join to tell us, we have to go into the first child to find it.  We don't want to have to be digging around in the possible first children, and children of first children, to find dot stars, not every time we call JoinExpr.simplify().

But this isn't the general case, it's a very simple one where the dot star is an immediate child to the join, and we create a lot of it, so it's worth taking a moment to detect and eliminate.