Small changes for brevity
(1) CorrespondExpr and RewriteExpr, as actually used, have very formulaic structures, and so their ids express information that's redundant in practice -- their input/output tape names, that the child of Correspond is a sequence, etc.  So I'm leveraging that to shorten their ids a bit.

(This structure might not 100% hold across every test we do, like we might try a weird thing in a correspond test or something, but those aren't the places we're needing VERBOSE_DEBUG right now.  ids are allowed to change depending on what we're actually looking for in the logs.)

Between the join-dot-star simplification in the previous step and this, ids are much more compact and now make more sense to me as a reader.  It looks from the logs like we're achieving our goal, at least for what we've tested so far.  The rule grammars are staying small and we're not going down garden paths.  (At least, not more than necessary.  We have to go down wrong paths sometimes because they might match later, we're just not going down those paths unnecessarily.)

(2) I put in an unconstrained generation test (well, partly constrained), but like replace/correspond above, the input/output tapes are very formulaic and it's annoying to wrangle a lot of "$i" and "$"o".  I added a field to the test, "io", that lets you specify the results as [input, output] string pairs instead.