Updated testGrammarMatch and testGrammarReplace for the new vocab scheme.
Removed a few tests along the way, and adjusted Counts and/or vocabs for some
to keep the amount of output generated to not be overwhelming.

Added a skipMatchOutputs option to testGrammar and testGenerate that we can add
(temporarily) if all we are interested in is finding out the number of outputs
generated. This can be helpful in test development.