BeginsWith rewrite rules
(1) Implemented rewrite rules where the context begins with "#" -- that is, that can only make their replacement at the beginnings of lines.

This ends up being pretty simple, it's basically just that instead of the continuation of the rewrite being "this", it's a matchDotStar.  Let us hope EndsWith rules are equally simple.

(2) While I was at it, I've been noticing that the different pieces of the rule skeleton call the inputs and outputs different names, and it was a little hard to remember when to change naming styles.  Replace, Match, and Correspond call them "from" and "to"; PreTape calls them "1" and "2"; ReplaceBlock and the tape system call them "input" and "output".  For brief periods of time we've even had XGrammar, X() params, XExpr, and constructX() params that would disagree for the same X.

"From" and "to" made sense because that's what they're called in the tabular syntax.  But at the same time, this is also what we call the tapes of Rename-like operations, and these *don't* line up, e.g. sometimes the fromTape from a rename perspective is the toTape from a replace perspective.

To avoid confusion between replace-family objects/functions and rename-family objects/functions, I decided to standardize the replace terms to "input" and "output" across the replace system.  (Except, of course, in the tabular syntax the programmer uses, they still type "from" and "to" like before; we need to keep that for compatibility.  This is just to keep the codebase more consistent.)