Adding endsWith rewrite rules
This commit adds beginsWith and endsWith rewrite rules.  It's not quite as simple as when there's only beginsWith; to get all four possibilities (normal, begin, ends, begins+ends) you end up needing to change three pieces of the equation.  All three are "continuations" in some sense:

  (a) The continuation of the deriv after matching the pattern
  (b) The continuation of the deriv after NOT matching the pattern
  (c) The input-tape continuation for the purposes of the negative clause

Because of this it's a bit convoluted to read the formulae off the code, although I'm not sure they're really convoluted enough to have 4 separate RewriteExpr variants rather than doing it like this.

Really it's only so tricky to read because the way we have to express formulae is so verbose.  Maybe after refactoring to a better way of writing derivs, this will look clearer.