Detecting/handling empty-input rules
(1) Empty-input rules -- rules where all of "pre", "input", and "post" can be epsilon -- require special detection and care.  If they're not constrained, they'll just lead to infinite generation because they can apply anywhere.  (For our current codebase, actually, they lead not just to infinite generation but an infinite loop and a blown stack!)

We haven't actually gotten around to allowing empty-input rules in a principled and constrained way, we've just been detecting them and eliminating them.  (With the new approach, I think it'll actually be quite a lot easier to handle them, so maybe that'll come soon, but not today.)

But our detection rule was still expecting them to be ReplaceGrammars rather than RewriteGrammars, so I fixed this.  Now they're detected and eliminated.

(2) There's a special corner case that our new rewrite approach wasn't quite catching.

We allow empty-input rules so long as they're beginsWith and/or endsWith (e.g. `R(eps->X|_#)`), because on some level they're not "really" empty, they're sensitive to word boundaries.  (We don't happen to implement word boundaries as special characters, but abstractly, it's a sort of "character" that's part of the input pattern we're trying to match.)

BeginsWith empty-input rules were working, but endsWith ones weren't.

To illustrate what's going wrong, think about a rule `R(eps->X|_#)` joined with an input string like "abc".  Run it to the end, where all of "abc" has been consumed and there's now {$i:abc, $o:abc}.  At this stage, everything is delta-able.  $i is delta-able because we're at the end of abc, the matched $o is delta-able because it matches $i and $i is nullable, and the rule itself is nullable because rules are effectively a sort of repeat.  So {$i:abc, $o:abc} passes the test for a valid output here... but the answer should be {$i:abc, $o:abcX}.  And there's no way to get that output -- we don't end up calling deriv on the rule because it's joined with the input "abc" and that's got no more material to send through the join.

The only thing that can possibly happen at this juncture is a delta call, the delta call that finishes off the "abc" input, so that's where we have to handle this.  So what we do now is check if the input material (pre/input/post) is nullable, and if it is, the delta of the rule is the output material (here, $o:X) rather than epsilon.

I think this might actually be (close to) the correct solution rather than a hack, I think it also gives us an avenue to handling the non-endsWith empty-input rules we're currently ignoring.  It's reminiscent of what we're doing with the extra delta term in Repeats, which makes sense because this is a special kind of repeat.